+++ Dockerfile
@@ -4,7 +4,6 @@ RUN apt update && apt -y install openssh-server git
 RUN mkdir /app
 WORKDIR /app
 COPY . .
-ENV GOPRIVATE=github.com/neutron-org/neutron
 RUN go mod download && \
     go build -a -o /go/bin/neutron_query_relayer ./cmd/neutron_query_relayer
 
@@ -15,4 +14,4 @@ ADD run.sh .
 COPY --from=builder /go/bin/neutron_query_relayer /bin/
 EXPOSE 9999
 
-ENTRYPOINT ["neutron_query_relayer"]
+ENTRYPOINT ["neutron_query_relayer", "start"]
